Made a separate class for status bar state control
