+++ api/src/opentrons/hardware_control/ot3api.py
@@ -142,7 +142,7 @@ from opentrons_hardware.hardware_control.motion_planning.move_utils import (
 
 from opentrons_hardware.firmware_bindings.constants import FirmwareTarget, USBTarget
 
-from opentrons_hardware.hardware_control import status_bar
+from .status_bar_state import StatusBarStateController
 
 mod_log = logging.getLogger(__name__)
 
@@ -211,7 +211,9 @@ class OT3API(
                 self._config.motion_settings, self._gantry_load
             )
         )
-        self._status_bar_state = StatusBarState.IDLE
+        self._status_bar_controller = StatusBarStateController(
+            self._backend.status_bar_interface()
+        )
 
         self._pipette_handler = OT3PipetteHandler({m: None for m in OT3Mount})
         self._gripper_handler = GripperHandler(gripper=None)
@@ -495,78 +497,8 @@ class OT3API(
             await asyncio.sleep(max(0, 0.25 - (now - then)))
         await self.set_lights(button=True)
 
-    async def _status_bar_idle(self) -> None:
-        self._status_bar_state = StatusBarState.IDLE
-        await self._backend.status_bar_controller().static_color(status_bar.WHITE)
-
-    async def _status_bar_running(self) -> None:
-        self._status_bar_state = StatusBarState.RUNNING
-        await self._backend.status_bar_controller().static_color(status_bar.BLUE)
-
-    async def _status_bar_paused(self) -> None:
-        self._status_bar_state = StatusBarState.PAUSED
-        await self._backend.status_bar_controller().pulse_color(status_bar.BLUE)
-
-    async def _status_bar_hardware_error(self) -> None:
-        self._status_bar_state = StatusBarState.HARDWARE_ERROR
-        await self._backend.status_bar_controller().flash_color(status_bar.RED)
-
-    async def _status_bar_software_error(self) -> None:
-        self._status_bar_state = StatusBarState.SOFTWARE_ERROR
-        await self._backend.status_bar_controller().static_color(status_bar.RED)
-
-    async def _status_bar_confirm(self) -> None:
-        # Confirm should revert to IDLE
-        self._status_bar_state = StatusBarState.IDLE
-        await self._backend.status_bar_controller().blink_once(
-            status_bar.GREEN, status_bar.WHITE
-        )
-
-    async def _status_bar_run_complete(self) -> None:
-        self._status_bar_state = StatusBarState.RUN_COMPLETED
-        await self._backend.status_bar_controller().static_color(status_bar.GREEN)
-
-    async def _status_bar_updating(self) -> None:
-        self._status_bar_state = StatusBarState.UPDATING
-        await self._backend.status_bar_controller().pulse_color(status_bar.WHITE)
-
-    async def _status_bar_activation(self) -> None:
-        # Activation should revert to IDLE
-        self._status_bar_state = StatusBarState.IDLE
-        await self._backend.status_bar_controller().blink_once(
-            status_bar.BLUE, status_bar.WHITE
-        )
-
-    async def _status_bar_disco(self) -> None:
-        # TODO - update implementation
-        colors = [
-            status_bar.GREEN,
-            status_bar.YELLOW,
-            status_bar.PURPLE,
-            status_bar.ORANGE,
-            status_bar.GREEN,
-            status_bar.YELLOW,
-            status_bar.PURPLE,
-            status_bar.ORANGE,
-            status_bar.WHITE,
-        ]
-        self._status_bar_state = StatusBarState.IDLE
-        await self._backend.status_bar_controller().cycle_colors(colors)
-
     async def set_status_bar_state(self, state: StatusBarState) -> None:
-        callbacks = {
-            StatusBarState.IDLE: self._status_bar_idle,
-            StatusBarState.RUNNING: self._status_bar_running,
-            StatusBarState.PAUSED: self._status_bar_paused,
-            StatusBarState.HARDWARE_ERROR: self._status_bar_hardware_error,
-            StatusBarState.SOFTWARE_ERROR: self._status_bar_software_error,
-            StatusBarState.CONFIRMATION: self._status_bar_confirm,
-            StatusBarState.RUN_COMPLETED: self._status_bar_run_complete,
-            StatusBarState.UPDATING: self._status_bar_updating,
-            StatusBarState.ACTIVATION: self._status_bar_activation,
-            StatusBarState.DISCO: self._status_bar_disco,
-        }
-        await callbacks[state]()
+        await self._status_bar_controller.set_status_bar_state(state)
 
     @ExecutionManagerProvider.wait_for_running
     async def delay(self, duration_s: float) -> None:
